# web-interface/backend/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CV Adapter Web Interface",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/api/generations/competences": {
      "post": {
        "tags": [
          "generations"
        ],
        "summary": "Generate Competences",
        "description": "Generate core competences from CV and job description.",
        "operationId": "generate_competences_v1_api_generations_competences_post",
        "parameters": [
          {
            "name": "language_code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Language Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateCompetencesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "title": "Response Generate Competences V1 Api Generations Competences Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/api/generations/cv": {
      "post": {
        "tags": [
          "generations"
        ],
        "summary": "Generate Cv",
        "description": "Generate a complete CV using core competences.",
        "operationId": "generate_cv_v1_api_generations_cv_post",
        "parameters": [
          {
            "name": "language_code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Language Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateCVRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/api/generations": {
      "get": {
        "tags": [
          "generations"
        ],
        "summary": "Get User Generations",
        "description": "Get all generated CVs for current user.",
        "operationId": "get_user_generations_v1_api_generations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeneratedCVResponse"
                  },
                  "type": "array",
                  "title": "Response Get User Generations V1 Api Generations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "generations"
        ],
        "summary": "Generate And Save Cv",
        "description": "Generate and save a new CV for job application.",
        "operationId": "generate_and_save_cv_v1_api_generations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratedCVCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedCVResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/v1/api/generations/{cv_id}": {
      "get": {
        "tags": [
          "generations"
        ],
        "summary": "Get Generated Cv",
        "description": "Get a specific generated CV.",
        "operationId": "get_generated_cv_v1_api_generations__cv_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cv Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedCVResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ContactRequest": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "value",
          "type"
        ],
        "title": "ContactRequest"
      },
...
      "GenerateCVRequest": {
        "properties": {
          "cv_text": {
            "type": "string",
            "title": "Cv Text"
          },
          "job_description": {
            "type": "string",
            "title": "Job Description"
          },
          "personal_info": {
            "$ref": "#/components/schemas/PersonalInfo"
          },
          "approved_competences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Approved Competences"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "cv_text",
          "job_description",
          "personal_info",
          "approved_competences"
        ],
        "title": "GenerateCVRequest"
      },
      "GenerateCompetencesRequest": {
        "properties": {
          "cv_text": {
            "type": "string",
            "title": "Cv Text"
          },
          "job_description": {
            "type": "string",
            "title": "Job Description"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "cv_text",
          "job_description"
        ],
        "title": "GenerateCompetencesRequest"
      },
      "GeneratedCVCreate": {
        "properties": {
          "language_code": {
            "type": "string",
            "title": "Language Code"
          },
          "content": {
            "type": "object",
            "title": "Content"
          },
          "detailed_cv_id": {
            "type": "integer",
            "title": "Detailed Cv Id"
          },
          "job_description_id": {
            "type": "integer",
            "title": "Job Description Id"
          }
        },
        "type": "object",
        "required": [
          "language_code",
          "content",
          "detailed_cv_id",
          "job_description_id"
        ],
        "title": "GeneratedCVCreate",
        "description": "Schema for creating a generated CV."
      },
      "GeneratedCVResponse": {
        "properties": {
          "language_code": {
            "type": "string",
            "title": "Language Code"
          },
          "content": {
            "type": "object",
            "title": "Content"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "detailed_cv_id": {
            "type": "integer",
            "title": "Detailed Cv Id"
          },
          "job_description_id": {
            "type": "integer",
            "title": "Job Description Id"
          }
        },
        "type": "object",
        "required": [
          "language_code",
          "content",
          "id",
          "created_at",
          "user_id",
          "detailed_cv_id",
          "job_description_id"
        ],
        "title": "GeneratedCVResponse",
        "description": "Schema for generated CV responses."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PersonalInfo": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "$ref": "#/components/schemas/ContactRequest"
          },
          "phone": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContactRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContactRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email"
        ],
        "title": "PersonalInfo"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "v1/auth/login"
          }
        }
      }
    }
  }
}